--- src/utils/DateUnitToMonth.js
@@ -1,8 +1,0 @@
-export const dateUnitToMonth = () => {
-
-    const standardStartDate = new Date(2023, 0, 1, 0, 0, 0)
-    const standardEndDate = new Date(2023, 11, 31, 23, 59, 59)
-
-    console.log(standardStartDate)
-    console.log(standardEndDate)
-}
